feat(hooks): hook to get profile stats
