+++ ui/hooks/src/index.ts
@@ -37,3 +37,4 @@ export {
 
 export { useShowFeedback } from './use-show-feedback';
 export { useTheme } from './use-theme';
+export { useProfileStats } from './use-profile-stats';
